metric/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=4.0.0=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39ha55989b_4
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.0=pyhd8ed1ab_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
babel=2.13.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
bleach=6.1.0=pyhd8ed1ab_0
brotli-python=1.1.0=py39h99910a6_1
bzip2=1.0.8=h8ffe710_4
ca-certificates=2023.7.22=h56e8100_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2022.12.7=pypi_0
cffi=1.16.0=py39ha55989b_0
charset-normalizer=2.1.1=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
debugpy=1.8.0=py39h99910a6_1
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.9.0=pypi_0
fqdn=1.5.1=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
imageio=2.31.5=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
importlib_resources=6.1.0=pyhd8ed1ab_0
intel-openmp=2023.2.0=h57928b3_50496
ipykernel=6.25.2=pyh60829e3_0
ipython=8.16.1=pyh5737063_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.1.1=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
json5=0.9.14=pyhd8ed1ab_0
jsonpointer=2.4=py39hcbf5309_3
jsonschema=4.19.1=pyhd8ed1ab_0
jsonschema-specifications=2023.7.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.19.1=pyhd8ed1ab_0
jupyter=1.0.0=pyhd8ed1ab_10
jupyter-lsp=2.2.0=pyhd8ed1ab_0
jupyter_client=8.4.0=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.4.0=py39hcbf5309_0
jupyter_events=0.8.0=pyhd8ed1ab_0
jupyter_server=2.8.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=4.0.7=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.25.0=pyhd8ed1ab_0
jupyterlab_widgets=3.0.9=pyhd8ed1ab_0
lazy-loader=0.3=pypi_0
libblas=3.9.0=19_win64_mkl
libcblas=3.9.0=19_win64_mkl
libffi=3.4.2=h8ffe710_5
libhwloc=2.9.3=default_haede6df_1009
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=19_win64_mkl
libsodium=1.0.18=h8d14728_1
libsqlite=3.43.2=hcfcfb64_0
libxml2=2.11.5=hc3477c8_1
libzlib=1.2.13=hcfcfb64_5
markupsafe=2.1.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=3.0.1=pyhd8ed1ab_0
mkl=2023.2.0=h6a75c08_50496
mpmath=1.3.0=pypi_0
nbclient=0.8.0=pyhd8ed1ab_0
nbconvert=7.9.2=pyhd8ed1ab_0
nbconvert-core=7.9.2=pyhd8ed1ab_0
nbconvert-pandoc=7.9.2=pyhd8ed1ab_0
nbformat=5.9.2=pyhd8ed1ab_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.0=pypi_0
notebook=7.0.6=pyhd8ed1ab_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.26.0=py39hddb5d58_0
openssl=3.1.3=hcfcfb64_0
overrides=7.4.0=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandoc=3.1.3=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=3.11.0=pyhd8ed1ab_0
prometheus_client=0.17.1=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py39ha55989b_1
pthreads-win32=2.9.1=hfa6e2cd_3
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.9.18=h4de0772_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.18.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python_abi=3.9=4_cp39
pytorch-fid=0.3.0=pypi_0
pytz=2023.3.post1=pyhd8ed1ab_0
pywin32=306=py39h99910a6_2
pywinpty=2.0.12=py39h99910a6_0
pyyaml=6.0.1=py39ha55989b_1
pyzmq=25.1.1=py39hea35a22_1
qtconsole-base=5.4.4=pyha770c72_0
qtpy=2.4.0=pyhd8ed1ab_0
referencing=0.30.2=pyhd8ed1ab_0
requests=2.28.1=pypi_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rpds-py=0.10.6=py39hf21820d_0
scikit-image=0.22.0=pypi_0
scipy=1.11.1=pypi_0
send2trash=1.8.2=pyh08f2357_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tbb=2021.10.0=h91493d7_1
terminado=0.15.0=py39hcbf5309_0
tifffile=2023.9.26=pypi_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.13=hcfcfb64_0
tomli=2.0.1=pyhd8ed1ab_0
torch=1.8.0+cu111=pypi_0
torchaudio=0.8.0=pypi_0
torchvision=0.9.0+cu111=pypi_0
tornado=6.3.3=py39ha55989b_1
traitlets=5.11.2=pyhd8ed1ab_0
types-python-dateutil=2.8.19.14=pyhd8ed1ab_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2023c=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=1.26.13=pypi_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
wcwidth=0.2.8=pyhd8ed1ab_0
webcolors=1.13=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.6.4=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
widgetsnbextension=4.0.9=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
winpty=0.4.3=4
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
